LayoutTests/sputnik/Conformance/07_Lexical_Conventions/7.9_Automatic_Semicolon_Insertion/S7.9_A6.3_T4-expected.txt

CONSOLE MESSAGE: line 81: SyntaxError: Expected token 'in'
S7.9_A6.3_T4

PASS Expected parsing failure

TEST COMPLETE
